Move embed filters to own sub namespace
