+++ src/Service/Embed/Filter/EmbedFilterInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Service\Embed;
+namespace App\Service\Embed\Filter;
 
 interface EmbedFilterInterface
 {
