+++ src/Service/Embed/Filter/YoutubeEmbedFilter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Service\Embed;
+namespace App\Service\Embed\Filter;
 
 class YoutubeEmbedFilter implements EmbedFilterInterface
 {
